+++ Com.Danliris.Service.Inventory.Mongo.Lib/MongoModels/InventoryMovementMongo.cs
@@ -25,11 +25,11 @@ namespace Com.Danliris.Service.Inventory.Mongo.Lib.MongoModels
 
         public string storageName { get; set; }
 
-        public int stockPlanning { get; set; }
+        public double stockPlanning { get; set; }
 
         public double before { get; set; }
 
-        public int quantity { get; set; }
+        public double quantity { get; set; }
 
         public double after { get; set; }
